moves `suppressWorkflowOutputs` from `Sanitizer` to `TaskTransformer` per @ahgittin comment
